Add watchdog monitoring scripts for PhidgetSBC
Watchdog is a linux daemon that can check the status of a system and
repair or reboot it if necessary. I use it on the PhidgetSBC to monitor
the controller_pws_main.py daemon. If that daemon crashes, it is
restarted. If that doesn’t work, the system is rebooted.