Config controller can be run without app.cfg
If app.cfg is missing when config controller is run (as it will be
during initial setup), then a skeleton config is loaded so that the app
still functions. This is only allowed during config setup as an
improperly configured data path, DB, or station list could be a bad
thing for other controllers.